% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasetempSeed.R
\docType{data}
\name{Data}
\alias{Data}
\title{This is data to be included in my package}
\format{
A data frame with 5 rows and 19 column
}
\usage{
data(Data)
}
\description{
This is data to be included in my package
}
\keyword{datasets}
