% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasetempSeed.R
\name{EstCoeff}
\alias{EstCoeff}
\title{Estimation of Base Temperature}
\usage{
EstCoeff(Data)
}
\arguments{
\item{Data}{Output TempInter Function}
}
\value{
\itemize{
  \item Coefficients: Estimate of Coefficients
}
}
\description{
Estimation of Base Temperature
}
\examples{
library("BaseTempSeed")
Inter_data <- TempInter(Data=Data,GPercentage=c(0.1,0.2, 0.3,0.4, 0.5,0.6, 0.7,.8,0.9))
Est<-EstCoeff(Inter_data)
}
\references{
\itemize{
\item Ellis, R. H., Simon, G., & Covell, S. (1987). The Influence of Temperature on Seed Germination Rate in Grain LegumesIII. A Comparison of Five Faba Bean Genotypes at Constant Temperatures Using a New Screening Method. Journal of Experimental Botany, 38(6), 1033–1043.
\item Garcia-Huidobro, J., Monteith, J. L., & Squire, G. R. (1982). Time, temperature and germination of pearl millet (Pennisetum typhoides S. & H.) I. Constant temperature. Journal of experimental botany, 33(2), 288-296.
}
}
